+++ my-dev-portal-api/package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "app.js",
   "scripts": {
+    "start-server": "node app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
